+++ packages/terra-data-grid/tests/wdio/data-grid-spec.js
@@ -270,6 +270,16 @@ Terra.describeViewports('DataGrid', ['medium', 'large'], () => {
       expect(browser.$('//*[@id="default-terra-data-grid-focusable-cell-table"]/tbody[2]/tr[1]/td[1]/div/button').isFocused()).toBe(true);
     });
 
+    it('selects the cell after pressing Escape on a button that was auto focused', () => {
+      browser.keys(['Tab', 'Tab', 'ArrowDown', 'ArrowRight', 'Escape']);
+      expect(browser.$('//*[@id="default-terra-data-grid-focusable-cell-table"]/tbody[2]/tr[1]/td[1]').isFocused()).toBe(true);
+    });
+
+    it('selects the button after pressing Enter on cell that has focus with an auto focusable element', () => {
+      browser.keys(['Tab', 'Tab', 'ArrowDown', 'ArrowRight', 'Escape', 'Enter']);
+      expect(browser.$('//*[@id="default-terra-data-grid-focusable-cell-table"]/tbody[2]/tr[1]/td[1]/div/button').isFocused()).toBe(true);
+    });
+
     it('clicks the button instead of diving in if it is the only component in the cell', () => {
       const modal = browser.$('[aria-modal="true"]');
       expect(modal.isDisplayed()).toBe(false);
@@ -311,6 +321,13 @@ Terra.describeViewports('DataGrid', ['medium', 'large'], () => {
       navigateToCell(2, 1);
       moveCurrentPositionBy(6, 0);
       Terra.validates.element('sticky-header-scroll', { selector: stickyHeaderSelector });
+    });
+
+    it('verifies that the table column header with action row is fixed when scrolling', () => {
+      browser.url('/raw/tests/cerner-terra-framework-docs/data-grid/data-grid/data-grid-sticky-header-with-actions');
+      navigateToCell(2, 1);
+      moveCurrentPositionBy(6, 0);
+      Terra.validates.element('sticky-header-with-actions-scroll', { selector: '#terra-data-grid-sticky-with-actions-table' });
     });
 
     it('validates that focused cells scroll into view when beneath sticky column header', () => {