+++ packages/terra-framework-docs/src/terra-dev-site/doc/aggregator/common/placeholder-list/PlaceholderListItem.jsx
@@ -24,7 +24,7 @@ const PlaceholderListItem = ({
       className={classNames}
       onClick={onSelect}
       onKeyDown={keyDown}
-      role="option"
+      role="tab"
     >
       {children}
     </li>