+++ packages/terra-framework-docs/src/terra-dev-site/doc/data-grid/WorklistDataGrid.1/Examples.3/WorklistDataGridFocusableCell.jsx
@@ -32,9 +32,9 @@ const WorklistDataGridFocusableCell = () => {
   const gridDataJSON = {
     cols: [
       { id: 'Column-0', displayName: 'Patient' },
-      { id: 'Column-1', displayName: 'Column 1' },
-      { id: 'Column-2', displayName: 'Column 2' },
-      { id: 'Column-3', displayName: 'Column 3' },
+      { id: 'Column-1', displayName: 'Action A' },
+      { id: 'Column-2', displayName: 'Action B' },
+      { id: 'Column-3', displayName: 'Action C' },
     ],
     rows: [
       {